Refactored COM types.
