+++ src/com/actions/object/addobject.ts
@@ -1,4 +1,4 @@
-import { COMEntity } from "../../types.ts";
+import { COMEntity } from "../../types";
 
 export default interface ADD_OBJECT {
     'PAYLOAD': COMEntity,
