+++ src/com/actions/object/deleteobject.ts
@@ -1,4 +1,4 @@
-import { COMEntity } from "../../types.ts";
+import { COMEntity } from "../../types";
 
 export default interface DELETE_OBJECT {
     'PAYLOAD': Partial<COMEntity> & { id: string },
