+++ src/com/actions/object/deselectobject.ts
@@ -1,4 +1,4 @@
-import { COMEntity } from "../../types.ts";
+import { COMEntity } from "../../types";
 
 export default interface DESELECT_OBJECT {
     'PAYLOAD': Partial<COMEntity> & { id: string },
