+++ src/com/actions/object/getallobjects.ts
@@ -1,4 +1,4 @@
-import { COMEntity } from "../../types.ts";
+import { COMEntity } from "../../types";
 
 export default interface GET_ALL_OBJECTS {
     'PAYLOAD': Map<string, COMEntity>,
