+++ src/com/actions/object/getobjects.ts
@@ -1,4 +1,4 @@
-import { COMEntity } from "../../types.ts";
+import { COMEntity } from "../../types";
 
 export default interface GET_OBJECTS {
     'PAYLOAD': { ids: string[] },
