+++ src/com/actions/object/selectobject.ts
@@ -1,4 +1,4 @@
-import { COMEntity } from "../../types.ts";
+import { COMEntity } from "../../types";
 
 export default interface SELECT_OBJECT {
     'PAYLOAD': Partial<COMEntity> & { id: string },
