+++ src/com/actions/object/updateobject.ts
@@ -1,4 +1,4 @@
-import { COMEntity } from "../../types.ts";
+import { COMEntity } from "../../types";
 
 export default interface UPDATE_OBJECT {
     'PAYLOAD': Partial<COMEntity> & { id: string },
