+++ src/com/types/COMBaseEntity.ts
@@ -1,0 +1,6 @@
+export type COMBaseEntity = {
+    id: string;
+    name: string;
+    entityType: 'pov' | 'light' | 'model' | 'primitive';
+    visible: boolean;
+}
