+++ src/com/types/COMEntity.ts
@@ -1,0 +1,6 @@
+import { type COMPov } from "./COMPov";
+import { type COMLight } from "./COMLight";
+import { type COMModel } from "./COMModel";
+import { type COMPrimitive } from "./COMPrimitive";
+
+export type COMEntity = COMPov | COMLight | COMModel | COMPrimitive;
