+++ src/com/types/COMGeometry.ts
@@ -1,0 +1,6 @@
+export type COMGeometry = {
+    name: string
+    width: number;
+    height: number;
+    depth: number;
+}
