+++ src/com/types/index.ts
@@ -1,0 +1,17 @@
+import { type COMEntity } from "./COMEntity";
+import { type COMPrimitive } from "./COMPrimitive";
+import { type COMModel } from "./COMModel";
+import { type COMLight } from "./COMLight";
+import { type COMPov } from "./COMPov";
+import { type COMGeometry } from "./COMGeometry";
+import { type COMMaterial } from "./COMMaterial";
+
+export type {
+    COMEntity,
+    COMPrimitive,
+    COMModel,
+    COMLight,
+    COMPov,
+    COMGeometry,
+    COMMaterial,
+}
